services/url-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
alembic==1.13.1
pytest==7.4.4
pytest-asyncio==0.23.3

# New dependencies
PyJWT==2.8.0
redis==5.0.1
slowapi==0.1.9
qrcode==7.4.2
Pillow==10.2.0

# Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
